+++ utilities/gcs/type_gcsevent.go
@@ -1,0 +1,52 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the 'License');
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an 'AS IS' BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package gcs
+
+import "time"
+
+// Event is the payload of a GCS event.
+type Event struct {
+	Kind                    string                 `json:"kind"`
+	ID                      string                 `json:"id"`
+	SelfLink                string                 `json:"selfLink"`
+	Name                    string                 `json:"name"`
+	Bucket                  string                 `json:"bucket"`
+	Generation              string                 `json:"generation"`
+	Metageneration          string                 `json:"metageneration"`
+	ContentType             string                 `json:"contentType"`
+	TimeCreated             time.Time              `json:"timeCreated"`
+	Updated                 time.Time              `json:"updated"`
+	TemporaryHold           bool                   `json:"temporaryHold"`
+	EventBasedHold          bool                   `json:"eventBasedHold"`
+	RetentionExpirationTime time.Time              `json:"retentionExpirationTime"`
+	StorageClass            string                 `json:"storageClass"`
+	TimeStorageClassUpdated time.Time              `json:"timeStorageClassUpdated"`
+	Size                    string                 `json:"size"`
+	MD5Hash                 string                 `json:"md5Hash"`
+	MediaLink               string                 `json:"mediaLink"`
+	ContentEncoding         string                 `json:"contentEncoding"`
+	ContentDisposition      string                 `json:"contentDisposition"`
+	CacheControl            string                 `json:"cacheControl"`
+	Metadata                map[string]interface{} `json:"metadata"`
+	CRC32C                  string                 `json:"crc32c"`
+	ComponentCount          int                    `json:"componentCount"`
+	Etag                    string                 `json:"etag"`
+	CustomerEncryption      struct {
+		EncryptionAlgorithm string `json:"encryptionAlgorithm"`
+		KeySha256           string `json:"keySha256"`
+	}
+	KMSKeyName    string `json:"kmsKeyName"`
+	ResourceState string `json:"resourceState"`
+}
